Module cleaver::fe

The .gib Syntax

Since there is only a single frontend at the moment the input language is defined by it. Head on over to the parser's source which holds the grammar. All of this is subject to change.

Modules

ast

error

parser

symbolizer

Structs

Position
Position represents a location within a certain source file. It is composed of a line- and column-number.

Functions

check
Run checks on parsed input. This includes type checks as well as semantic checks.

check_with_diag

parse
Runs the parser and symbolizer on all given files, returning an AST for each of them.

parse_with_diag