Struct cleaver::front::ast::Node

pub struct Node<I> {
    pub pos: Position,
    pub node: I,
}

Fields
pos: Position node: I
Methods

impl Node<Statement>

fn visit_stmt<'a>(&'a self, do_stmt: &mut FnMut(&'a Node<Statement>) -> bool) -> bool

Visit all statements with a given closure.

The given closure is called for each statement with a reference to the given node. This reference inherits the lifetime of self. The return value indicates whether further statements should be visited or not.

Use this method as alternative to the Visitor.

fn visit_expr<'a>(&'a self, do_expr: &mut FnMut(&'a Node<Expression>) -> bool) -> bool

Visit all expressions with a given closure.

Same as visit_stmt, but for expressions.

impl Node<Expression>

fn visit_expr<'a>(&'a self, do_expr: &mut FnMut(&'a Node<Expression>) -> bool) -> bool

Visit all expressions with a given closure.

Same as visit_stmt, but for expressions.