Module cleaver::utils::lenny

Lenny Face Generator (づσ.σ)づ .

This module maintains a set of hardcoded components for Lenny Faces. Generate as many of them as you need.

Source: https://github.com/Twentysix26/Red-Cogs

Functions

fliptable
(╯°□°）╯︵ ┻━┻

gen
Generate a new Lenny Face.

pylons

The Overmind has come to destroy all that we hold dear and assimilate us into itself. And I say to thee, this shall not come to pass! Aiur shall not fall! Executor, I stand ready!